requirements/base.txt
Django==1.6.5
South==1.0.2
django-allauth==0.19.0
django-angular==0.7.10
